update PROGRAMS test
