+++ src/data/types/program.ts
@@ -1,0 +1,23 @@
+export const PROGRAM_SCHEMA = {
+  $id: 'Program.schema.json',
+  type: 'object',
+  properties: {
+    name: {
+      $ref: 'LocalizableString',
+    },
+    url: {
+      $ref: 'LocalizableString',
+    },
+  },
+  required: ['name', 'url'],
+} as const;
+
+export const PROGRAMS_SCHEMA = {
+  $id: 'Programs',
+  title: 'Programs',
+  type: 'object',
+  patternProperties: {
+    '^.*$': { $ref: 'Program.schema.json' },
+  },
+  additionalProperties: false,
+} as const;
